requirements.txt
python-telegram-bot==20.6
telethon==1.32.1
python-dotenv==1.0.1
requests==2.31.0